requirements.txt
openai>=2.7.0
reportlab>=4.2.0
fastapi>=0.116.0
pydantic>=2.11.0
uvicorn>=0.35.0
opentelemetry-api>=1.31.1
opentelemetry-sdk>=1.31.1
opentelemetry-exporter-otlp>=1.31.1
